+++ modules/presets/preset.js
@@ -17,6 +17,8 @@ export function presetPreset(presetID, preset, addable, allFields, allPresets) {
   let _resolvedMoreFields;  // cache
   let _searchName; // cache
   let _searchNameStripped; // cache
+  let _searchAliases; // cache
+  let _searchAliasesStripped; // cache
 
   _this.id = presetID;
 
@@ -25,6 +27,8 @@ export function presetPreset(presetID, preset, addable, allFields, allPresets) {
   _this.originalTerms = (_this.terms || []).join();
 
   _this.originalName = _this.name || '';
+
+  _this.originalAliases = _this.aliases || '';
 
   _this.originalScore = _this.matchScore || 1;
 
@@ -123,6 +127,11 @@ export function presetPreset(presetID, preset, addable, allFields, allPresets) {
   };
 
 
+  _this.aliases = () => {
+    return _this.t('aliases', { 'default': _this.originalAliases }).trim().split(/\s*\n\s*/);
+  };
+
+
   _this.terms = () => _this.t('terms', { 'default': _this.originalTerms })
     .toLowerCase().trim().split(/\s*,+\s*/);
 
@@ -135,13 +144,24 @@ export function presetPreset(presetID, preset, addable, allFields, allPresets) {
 
   _this.searchNameStripped = () => {
     if (!_searchNameStripped) {
-      _searchNameStripped = _this.searchName();
-      // split combined diacritical characters into their parts
-      if (_searchNameStripped.normalize) _searchNameStripped = _searchNameStripped.normalize('NFD');
-      // remove diacritics
-      _searchNameStripped = _searchNameStripped.replace(/[\u0300-\u036f]/g, '');
+      _searchNameStripped = stripDiacritics(_this.searchName());
     }
     return _searchNameStripped;
+  };
+
+  _this.searchAliases = () => {
+    if (!_searchAliases) {
+      _searchAliases = _this.aliases().map(alias => alias.toLowerCase());
+    }
+    return _searchAliases;
+  };
+
+  _this.searchAliasesStripped = () => {
+    if (!_searchAliasesStripped) {
+      _searchAliasesStripped = _this.searchAliases();
+      _searchAliasesStripped = _searchAliasesStripped.map(stripDiacritics);
+    }
+    return _searchAliasesStripped;
   };
 
   _this.isFallback = () => {
@@ -305,6 +325,14 @@ export function presetPreset(presetID, preset, addable, allFields, allPresets) {
     }
   }
 
+
+  function stripDiacritics(s) {
+    // split combined diacritical characters into their parts
+    if (s.normalize) s = s.normalize('NFD');
+    // remove diacritics
+    s = s.replace(/[\u0300-\u036f]/g, '');
+    return s;
+  }
 
   return _this;
 }