+++ CHANGELOG.md
@@ -60,12 +60,14 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 * Imply `access=no` in access field of `highway=construction` objects ([#9102])
 * Don't show non-language tag-suffixes in multilingual name field ([#9124], thanks [@wcedmisten])
 * Render horse riding centers like farmyards ([#9118])
+* Support searching presets by their `aliases` ([#6139])
 #### Other
 * Redact more API tokens from custom imagery sources in changeset metadata tags ([#8976], thanks [@k-yle])
 #### :hammer: Development
 * Switch build system to [esbuild](https://esbuild.github.io/) for much faster builds ([#8774], thanks [@mbrzakovic] and [@bhousel])
 * Upgrade some dependencies: maki to `v7.1`, `fontawesome` to `v6.1`, `d3` to `v7.3`, `node-diff` to `v3.1`, `mocha` to `v9.2`, `svg-sprite` to `v1.5.4`, `marked` to `v4.0`
 
+[#6139]: https://github.com/openstreetmap/iD/issues/6139
 [#8774]: https://github.com/openstreetmap/iD/pull/8774
 [#8811]: https://github.com/openstreetmap/iD/issues/8811
 [#8905]: https://github.com/openstreetmap/iD/issues/8905